+++ agnocomplete/widgets.py
@@ -6,7 +6,7 @@ from distutils.version import StrictVersion
 from django import get_version
 
 from django.forms import widgets
-from django.core.urlresolvers import reverse_lazy
+from django.urls import reverse_lazy
 from django.utils.encoding import force_text as text
 from django.conf import settings